requirements.txt
requests==2.23.0
python-dotenv==0.13.0
dialogflow==1.0.0
protobuf==3.12.2
